chore: 共通コンポーネントにあるSelectを使わないようにした
Why: ラベルやメッセージはフォーム専用のため、共通コンポーネントに含めると再利用性が低くなる。そこでSelectからBaseFieldを外した結果、共通コンポーネントとしての意味がなくなり、削除した。
