--- apps/web/src/components/inputs/Select/Select.module.css
@@ -1,3 +1,0 @@
-.select-trigger {
-  min-width: 200px;
-}
