--- apps/web/src/components/inputs/Select/index.ts
@@ -1,1 +1,0 @@
-export { Select, SelectItem } from "./Select"
